ALL password domins ADD RSA and SHA512
from web broswer,we encrypt 【password+64bit salt】 with 1024-RSA public pem,and send the encryption_data to server.
in the server,we get original password,add a (different ) 64-bit salt,then sha512,storge SHA in the database.
we do not use TWO_fish_AES or TWO_sha_512 or bcrypt_SHA512 because the VPS is fucking slow.