password strength test function
when a new user register his/her account,we can test the strength of the password and storge the strength_code into database.
in webpage:【user center】 we can score the safety of a account by judging the strength of his/her password/passphrase.
 It's one of five values: VERY_WEAK, WEAK, REASONABLE, STRONG, and VERY_STRONG.
we recommend a user can use passphrase rather than password.